+++ src/main/java/com/geekbrains/domain/Task.java
@@ -1,4 +1,4 @@
-package com.geekbrains;
+package com.geekbrains.domain;
 
 
 import org.springframework.format.annotation.DateTimeFormat;
