+++ src/main/java/com/geekbrains/repo/TaskRepository.java
@@ -1,6 +1,6 @@
 package com.geekbrains.repo;
 
-import com.geekbrains.Task;
+import com.geekbrains.domain.Task;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
